Add experimental support for indenting paste content with tab
This breaks numerous things like:

* undo behavior once something has been indented
* skipping to the submit button of course

Additionally it doesn't support shift+tab as reversal of indentation
yet. Still, for my personal taste it's better than nothing, it was
driving me crazy.

Possibly this will end up with using some sane 3rd party package for
this case but this was a nice little experiment for a Sunday night ^^