Don't normalize paste's language too soon
In other words, if user entered "json" as a language we don't have to
necessarily normalize it to "js" since "json" is an alias recognized by
KeyLighter, too.

Now we actually store the original user's language as they provided it,
and we only _render_ the language as "plaintext" in case it is not
recognized by KeyLighter (yet?). This should provide nice forwards
compatibility.