Add naive spam prevention tailored to my personal needs
Adds a very opinionated service for detecting That One Particular™
paste pattern they love adding on paste.sobak.pl. As silly as it
seems, this may actually do a trick, especially considering that
they post M3U playlist files which are expected to be in a particular
format.

I kinda hope that they'll just choose different service at this point